Rename .lift to .lifting to avoid the lift() method shadowing the lift module. (#127)
* Rename .lift to .lifting to avoid the lift() method shadowing the lift module.

* Tidy the code and make future development easier.

- Remove all absolute import of pyvex.
- Remove all absolute import of Python modules that should be relative
  imports.
- Split lifting.__init__.py into separate files.
- Delay an import from angr to avoid circular import.
