+++ pyvex/lifting/post_processor.py
@@ -1,0 +1,12 @@
+
+class Postprocessor(object):
+    def __init__(self, irsb):
+        self.irsb = irsb
+
+    def postprocess(self):
+        """
+        Modify the irsb
+
+        All of the postprocessors will be used in the order that they are registered
+        """
+        pass
